+++ src/main/java/WoWSSSC/model/shipprofile/profile/Fighters.java
@@ -10,15 +10,15 @@ import lombok.Data;
 public class Fighters
 {
     private long avg_damage;
-    private long cruise_speed;
+    private float cruise_speed;
     private long fighters_id;
     private String fighters_id_str;
-    private long gunner_damage;
+    private float gunner_damage;
     private long max_ammo;
-    private long max_health;
+    private float max_health;
     private String name;
     private long plane_level;
-    private long prepare_time;
+    private float prepare_time;
     private long squadrons;
     private Fighters_Count_In_Squadron count_in_squadron;
 }
